organise file structure and modularise schemas/resolvers
