+++ package.json
@@ -18,6 +18,7 @@
     "express": "^4.16.3",
     "graphql": "^0.13.2",
     "graphql-tools": "^3.0.4",
+    "lodash": "^4.17.10",
     "mongoose": "^5.1.7",
     "passport": "^0.4.0",
     "passport-localapikey": "^0.0.3"
